% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilinearGrad.R
\name{bilinearGrad}
\alias{bilinearGrad}
\title{Gradient for bilinear interpolation}
\usage{
bilinearGrad(loc, cov_list)
}
\arguments{
\item{loc}{Point at which the gradient should be evaluated (2d vector)}

\item{cov_list}{List of J (number of covariates) "raster like" elements.
A raster like element is a 3 elements list with named elements
1) "x" a vector of increasing x locations (at which the covariate is sampled)
2) "y" a vector of increasing y locations (at which the covariate is sampled)
3) "z" a size(x)*size(y) matrix giving covariate values at location (x, y)}
}
\value{
Gradient of all covariates at the given location
}
\description{
Gradient for bilinear interpolation
}
