% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilinearGrad.R
\name{bilinearGradArray}
\alias{bilinearGradArray}
\title{Apply bilinearGrad to matrix of locations}
\usage{
bilinearGradArray(locs, cov_list)
}
\arguments{
\item{locs}{Matrix of locations (with two columns)}

\item{cov_list}{List of J (number of covariates) "raster like" elements.
A raster like element is a 3 elements list with named elements
1) "x" a vector of increasing x locations (at which the covariate is sampled)
2) "y" a vector of increasing y locations (at which the covariate is sampled)
3) "z" a size(x)*size(y) matrix giving covariate values at location (x, y)}
}
\value{
An array of dimension (n, 2, J), where n is the number of locations,
and J is the number of covariates.
}
\description{
Apply bilinearGrad to matrix of locations
}
