% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rhabit.R
\name{coef.rhabit}
\alias{coef.rhabit}
\title{Extract the coefficients associated with the covariates in a rhabit model}
\usage{
\method{coef}{rhabit}(object, complete = TRUE, ...)
}
\arguments{
\item{object}{a rhabit model fit using fit_langevin_ud}

\item{complete}{argument specifies whether  NA should be displayed or not. its value is set to TRUE by default}

\item{...}{additional arguments.}
}
\value{
the coefficients of the fitted rhabit  model
}
\description{
Extract the coefficients associated with the covariates in a rhabit model
}
\examples{
data(tracks)
fitted_langevin <- fit_langevin_ud(cbind(x,y) ~ grad_c1, data = tracks)
coef(fitted_langevin)
}
