% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{covariates}
\alias{covariates}
\title{This is the description of the data included in package Rhabit.}
\description{
covariates is a set of two covariates, simulated for the  RLangevin_MM vignette, included in the paxkage to help to reproduce the vgnette code
}
\references{
\url{data_blah.com}
}
\keyword{data}
