% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientInterpolation.R
\name{covGradAtLocs}
\alias{covGradAtLocs}
\title{Gradient of covariate field}
\usage{
covGradAtLocs(locs, xgrid, ygrid, cov_array, lag_inter = 10,
  grad_fun = NULL)
}
\arguments{
\item{locs}{Matrix of locations where the gradient should be evaluated}

\item{xgrid}{Grid on which the covariates are known}

\item{ygrid}{Grid on which the covariates are known}

\item{cov_array}{Array of values of the covariates at the points given by
xgrid and ygrid, of dimensions (length(xgrid),length(ygrid),length(beta)).}

\item{lag_inter}{integer specifying which points of cov_array will be used
for the interpolation, default to 10}

\item{grad_fun}{Optional list of functions taking a 2d vector and returning
a two 2d vector for the gradient}
}
\value{
Three-dimensional array of gradients of covariate fields.
The rows index time, the columns are the dimensions (x and y),
and the layers index the covariates.
}
\description{
Gradient of covariate field
}
