% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinUD.R
\name{fit_langevin_ud}
\alias{fit_langevin_ud}
\title{Obtaining UD estimate of an animal using Langevin movement model}
\usage{
fit_langevin_ud(
  formula,
  data,
  with_speed = TRUE,
  alpha = 0.95,
  leverage = FALSE
)
}
\arguments{
\item{formula}{a formula of the form cbind(x,y)~  V1 + V2}

\item{data}{a data frame with columns x, y and V1_x, V1_y, V2_x, V2_y
as well as a times column.}

\item{with_speed}{Logical. If TRUE, the speed parameter is estimated
Other wise it is set to one}

\item{alpha}{Confidence level (default: 0.95, i.e. 95\% confidence intervals)}

\item{leverage}{Logical. If TRUE, the standardized residuals and the leverage
are returned. Default: FALSE. Might not work when there are many
 observations,  because it creates an n times n matrix.}
}
\value{
A list of: est, the vector of estimates, and var, the
covariance matrix of the estimates.
}
\description{
Obtaining UD estimate of an animal using Langevin movement model
}
