% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getUD.R
\name{getUD}
\alias{getUD}
\title{Obtaining the classical RSF UD}
\usage{
getUD(covariates, beta, log = F)
}
\arguments{
\item{covariates}{covariates list, that must be a list of "raster-like"
elements. It is for the moment supposed that the grid is regular and
shared by all elemebts of covariates}

\item{beta}{beta values, must be the same length as covariates}

\item{log}{logical: get log value? by default = F}
}
\value{
a raster-like list
}
\description{
Obtaining the classical RSF UD
}
