% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientInterpolation.R
\name{gradLogUD}
\alias{gradLogUD}
\title{Gradient of the log of the utilisation distribution}
\usage{
gradLogUD(beta, loc, xgrid, ygrid, cov_array, grad_fun = NULL)
}
\arguments{
\item{beta}{Vector of resource selection coefficients}

\item{loc}{Point at which the gradient should be evaluated (2d vector)}

\item{xgrid}{Grid on which the covariates are known}

\item{ygrid}{Grid on which the covariates are known}

\item{cov_array}{Array of values of the covariates at the points given by
xgrid and ygrid, of dimensions (length(xgrid),length(ygrid),length(beta)).}

\item{grad_fun}{Optional list of functions taking a 2d vector and returning
a two 2d vector for the gradient}
}
\value{
Gradient of the log-UD in loc.
}
\description{
Gradient of the log of the utilisation distribution
}
