% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bilinearGrad.R
\name{gridCell}
\alias{gridCell}
\title{Extract covariate grid cell}
\usage{
gridCell(loc, xgrid, ygrid, covmat)
}
\arguments{
\item{loc}{Location}

\item{xgrid}{Grid along x axis}

\item{ygrid}{Grid along y axis}

\item{covmat}{Covariate values on (xgrid,ygrid)}
}
\value{
List of two elements: coords (coordinates of bottom-left and
top-right corners of the grid cell), and values (values of the covariate 
at the four corners of the grid cell).
}
\description{
Extract covariate grid cell
}
