% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/in_bbox.R
\name{in_bbox}
\alias{in_bbox}
\title{Check if a point is in the bouding box}
\usage{
in_bbox(point, bbox)
}
\arguments{
\item{point}{The point to be checked a c(x,y) vector}

\item{bbox}{The bounding box bbox[i,j] if i=1 is the minimal value for the bbox and i=2 the maximal value, j=1 corresponds to the x dimension and j=2 to the y dimension}
}
\value{
a boolean
}
\description{
Check if a point is in the bouding box
}
