% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinUD.R
\name{langevinUD}
\alias{langevinUD}
\title{Obtaining UD estimate of an animal using Langevin movement model}
\usage{
langevinUD(
  locs,
  times,
  ID = NULL,
  grad_array,
  with_speed = TRUE,
  alpha = 0.95,
  leverage = FALSE
)
}
\arguments{
\item{locs}{Matrix of locations having size, say, n}

\item{times}{Vector of observation times, must be of length n}

\item{ID}{Vector of track identifiers, must be of length n}

\item{grad_array}{Array of gradients of covariates, evaluated at
the locations of the observations, must be 3d array of dim(n, 2, J)
where J is the amount of covariates}

\item{with_speed}{Logical. If TRUE, the speed parameter is estimated
Other wise it is set to one}

\item{alpha}{Confidence level (default: 0.95, i.e. 95\% confidence intervals)}

\item{leverage}{Logical. If TRUE, the standardized residuals and the leverage
are returned. Default: FALSE. Might not work when there are many observations,
because it creates an n times n matrix.}
}
\value{
A list of: est, the vector of estimates, and var, the
covariance matrix of the estimates.
}
\description{
Obtaining UD estimate of an animal using Langevin movement model
}
