% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMALA.R
\name{logRSFinterp}
\alias{logRSFinterp}
\title{Evaluate the log-RSF based on interpolated covariates
(used in simMALA)}
\usage{
logRSFinterp(locs, beta, cov_list)
}
\arguments{
\item{locs}{Matrix of locations where the log-RSF should be evaluated}

\item{beta}{Vector of resource selection coefficients}

\item{cov_list}{List of J (number of covariates) "raster like" elements.
A raster like element is a 3 elements list with named elements
1) "x" a vector of increasing x locations (at which the covariate is sampled)
2) "y" a vector of increasing y locations (at which the covariate is sampled)
3) "z" a size(x)*size(y) matrix giving covariate values at location (x, y)}
}
\description{
Evaluate the log-RSF based on interpolated covariates
(used in simMALA)
}
