% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nllkEuler.R
\name{nllkEuler}
\alias{nllkEuler}
\title{Negative log-likelihood function for Langevin movement model
under Euler discretization}
\usage{
nllkEuler(beta, gamma2 = 1, locs, times, ID = NULL, grad_array)
}
\arguments{
\item{beta}{Parameters of the utilisation distribution}

\item{gamma2}{Diffusion parameter}

\item{locs}{Matrix of observed locations (two columns: x, y)}

\item{times}{Vector of times of observations}

\item{ID}{Vector of track identifiers, must be of length n}

\item{grad_array}{Three-dimensional array of gradients of covariate fields. 
The rows index time, the columns are the dimensions (x and y), and the 
layers index the covariates.}
}
\value{
Negative log-likelihood
}
\description{
Negative log-likelihood function for Langevin movement model
under Euler discretization
}
