% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rhabit.R
\name{print.rhabit}
\alias{print.rhabit}
\title{Print the formula used to specify the model and the estimated coefficients}
\usage{
\method{print}{rhabit}(x, ...)
}
\arguments{
\item{x}{a rhabit model fit using fit_langevin_ud}

\item{...}{additionnal arguments, might be digits a non-null value specifying the minimum number 
of digits to be printed in values.}
}
\value{
The rhabit object passed in argument is returned.
}
\description{
Print the formula used to specify the model and the estimated coefficients
}
\examples{
data(tracks)
fitted_langevin <- fit_langevin_ud(cbind(x,y) ~ grad_c1, data = tracks)
print(fitted_langevin)
}
