% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpatialCov.R
\name{rasterToDataFrame}
\alias{rasterToDataFrame}
\title{Transform (x,y,z) list to data.frame}
\usage{
rasterToDataFrame(my_list, level = NULL)
}
\arguments{
\item{my_list}{A raster like list containing elements x, y, z
(see simSpatialCov)}

\item{level}{An optional factor argument to add to a level column
(useful to pu color for instance)}
}
\value{
A data.frame (convenient for ggplot2)
}
\description{
Transform (x,y,z) list to data.frame
}
