% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rasterToRhabit.R
\name{rasterToRhabit}
\alias{rasterToRhabit}
\title{Raster to Rhabit}
\usage{
rasterToRhabit(r)
}
\arguments{
\item{r}{Raster object}
}
\value{
List of three elements: x (grid of x values), y (grid of y values),
and z (matrix of values)
}
\description{
Transform a raster into a (x,y,z) list, as used in Rhabit
}
