% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rhabit-package.R
\docType{package}
\name{Rhabit-package}
\alias{Rhabit}
\alias{Rhabit-package}
\title{Rhabit: Estimation of animal habitat selection using the Langevin movement model}
\description{
Estimation of animal habitat selection using the Langevin movement model.
}
\author{
\strong{Maintainer}: Pierre Gloaguen \email{pierre.gloaguen@agroparistech.fr}

Authors:
\itemize{
  \item Théo Michelot \email{tm75@st-andrews.ac.uk}
}

Other contributors:
\itemize{
  \item Marie-Pierre Etienne \email{marie-pierre.etienne@agrocampus-ouest.fr} [contributor]
}

}
\keyword{internal}
