% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rhabitToRaster.R
\name{rhabitToRaster}
\alias{rhabitToRaster}
\title{Rhabit to raster}
\usage{
rhabitToRaster(r)
}
\arguments{
\item{r}{List of three elements: x (grid of x values), y (grid of y values),
and z (matrix of values)}
}
\value{
Raster object
}
\description{
Transform a (x,y,z) list, as used in Rhabit, into a raster object
}
