% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simLangevinMM.R
\name{simLangevinMM}
\alias{simLangevinMM}
\title{Simulate from the Langevin movement model}
\usage{
simLangevinMM(beta, gamma2 = 1, times, loc0, xgrid, ygrid, cov_array,
  lag_inter = 10, grad_fun = NULL)
}
\arguments{
\item{beta}{Vector of resource selection coefficients}

\item{gamma2}{Scalar speed parameter}

\item{times}{Vector of times of observations. Should have a small
time step between each, such that the Euler scheme is valid}

\item{loc0}{Vector of coordinates (x,y) of initial location}

\item{xgrid}{Grid on which the covariates are known}

\item{ygrid}{Grid on which the covariates are known}

\item{cov_array}{Array of values of the covariates at the points given by}

\item{lag_inter}{integer specifying which points of cov_array will be used
for the interpolation, default to 10
xgrid and ygrid, of dimensions (length(xgrid),length(ygrid),length(beta)).}

\item{grad_fun}{Optional list of functions taking a 2d vector and returning
a two 2d vector for the gradient}
}
\description{
This function is based on the Euler approximation.
}
