% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simMALA.R
\name{simMALA}
\alias{simMALA}
\title{Simulate from Metropolis-adjusted Langevin algorithm}
\usage{
simMALA(beta, gamma2 = 1, times, loc0, cov_list = NULL,
  grad_fun = NULL, silent = FALSE)
}
\arguments{
\item{beta}{Vector of resource selection coefficients}

\item{gamma2}{Scalar speed parameter}

\item{times}{Vector of times of observations. Should have a small
time step between each, such that the Euler scheme is valid}

\item{loc0}{Vector of coordinates (x,y) of initial location}

\item{cov_list}{List of J (number of covariates) "raster like" elements.
A raster like element is a 3 elements list with named elements
1) "x" a vector of increasing x locations (at which the covariate is sampled)
2) "y" a vector of increasing y locations (at which the covariate is sampled)
3) "z" a size(x)*size(y) matrix giving covariate values at location (x, y)}

\item{grad_fun}{Optional list of functions taking a 2d vector and returning
a two 2d vector for the gradient}

\item{silent}{logical, should simulation advancement be shown?}
}
\description{
Simulate from Metropolis-adjusted Langevin algorithm
}
