% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simSpatialCov.R
\name{simSpatialCov}
\alias{simSpatialCov}
\title{Simulate random covariate field}
\usage{
simSpatialCov(lim, nu, rho, sigma2, resol = 1, mean_function = NULL,
  raster_like = FALSE)
}
\arguments{
\item{lim}{Vector of limits of the map (xmin, xmax, ymin, ymax)}

\item{nu}{smoothness parameter}

\item{rho}{range parameter}

\item{sigma2}{variance parameter}

\item{resol}{Grid resolution (defaults to 1)}

\item{mean_function}{function giving mean of a spatial coordinate. Must
be coded to take a 2 (or 3)d vector and return a scalar. By default set
to return 0 for every coordinates}

\item{raster_like}{logical specifying the return output. If TRUE, then
a list with xgrid, ygrid and  matrix is returned convenient for image.plot.
If FALSE a data.frame is returned, which might be more convenient for ggplot.
Default to FALSE.}
}
\value{
either a data.frame (convenient for ggplot2) or a raster like list
}
\description{
Simulate random covariate field
}
