% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rhabit.R
\name{speed_coef}
\alias{speed_coef}
\title{Show the estimated speed of the Langevin model}
\usage{
speed_coef(x)
}
\arguments{
\item{x}{a rhabit object obtained by fit_langevin_ud}
}
\value{

}
\description{
Show the estimated speed of the Langevin model
}
\examples{
data(tracks)
fitted_langevin <- fit_langevin_ud(cbind(x,y) ~ grad_c1, data = tracks)
speed_coef(fitted_langevin)
}
