% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tracks}
\alias{tracks}
\title{tracks is a set of simulated tracks based on a Langevin model with  two covariates, simulated for the  RLangevin_MM vignette. The code is also available in the vignette.}
\description{
tracks is a set of simulated tracks based on a Langevin model with  two covariates, simulated for the  RLangevin_MM vignette. The code is also available in the vignette.
}
\references{
\url{data_blah.com}
}
\keyword{data}
