% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AICEuler.R
\name{AICEuler}
\alias{AICEuler}
\title{Akaike Information Criterion for Langevin movement model
under Euler discretization}
\usage{
AICEuler(beta, gamma2 = 1, locs, times, ID = NULL, grad_list)
}
\arguments{
\item{beta}{Parameters of the utilisation distribution}

\item{gamma2}{Diffusion parameter}

\item{locs}{Matrix of observed locations (two columns: x, y)}

\item{times}{Vector of times of observations}

\item{ID}{Vector of track identifiers, must be of length n}

\item{grad_list}{list of gradients of covariates, evaluated at
the locations of the observations, must be 3d array of dim(n, 2, J)
where J is the amount of covariates .}
}
\value{
AIC
}
\description{
Akaike Information Criterion for Langevin movement model
under Euler discretization
}
