% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientInterpolation.R
\name{covGradAtLocs}
\alias{covGradAtLocs}
\title{Gradient of covariate field}
\usage{
covGradAtLocs(locs, cov_list = NULL, grad_fun = NULL)
}
\arguments{
\item{locs}{Matrix of locations where the gradient should be evaluated}

\item{cov_list}{List of J (number of covariates) "raster like" elements.
A raster like element is a 3 elements list with named elements
1) "x" a vector of increasing x locations (at which the covariate is sampled)
2) "y" a vector of increasing y locations (at which the covariate is sampled)
3) "z" a size(x)*size(y) matrix giving covariate values at location (x, y)}

\item{grad_fun}{Optional list of functions taking a 2d vector and returning
a two 2d vector for the gradient}
}
\value{
Three-dimensional array of gradients of covariate fields.
The rows index time, the columns are the dimensions (x and y),
and the layers index the covariates.
}
\description{
Gradient of covariate field
}
