% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gradientInterpolation.R
\name{gradLogUD}
\alias{gradLogUD}
\title{Gradient of the log of the utilisation distribution}
\usage{
gradLogUD(beta, loc, cov_list = NULL, grad_fun = NULL, check = T)
}
\arguments{
\item{beta}{Vector of resource selection coefficients}

\item{loc}{Point at which the gradient should be evaluated (2d vector)}

\item{cov_list}{List of J (number of covariates) "raster like" elements.
A raster like element is a 3 elements list with named elements
1) "x" a vector of increasing x locations (at which the covariate is sampled)
2) "y" a vector of increasing y locations (at which the covariate is sampled)
3) "z" a size(x)*size(y) matrix giving covariate values at location (x, y)}

\item{grad_fun}{Optional list of functions taking a 2d vector and returning
a two 2d vector for the gradient}
}
\value{
Gradient of the log-UD in loc.
}
\description{
Gradient of the log of the utilisation distribution
}
