% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinUD.R
\name{langevinUD}
\alias{langevinUD}
\title{Obtaining UD estimate of an animal using Langevin movement model}
\usage{
langevinUD(locs, times, ID = NULL, gradsArray, withspeed = TRUE)
}
\arguments{
\item{locs}{Matrix of locations}

\item{times}{Vector of observation times}

\item{ID}{Vector of track identifiers}

\item{gradsArray}{Array of gradients of covariates, evaluated at
the locations of the observations}

\item{withspeed}{Logical. If TRUE, the speed parameter is estimated
Other wise it is set to one}
}
\value{
A list of: est, the vector of estimates, and var, the
covariance matrix of the estimates.
}
\description{
Obtaining UD estimate of an animal using Langevin movement model
}
