% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCovariates.R
\name{plotCovariates}
\alias{plotCovariates}
\title{Plotting the classical RSF UD}
\usage{
plotCovariates(covariates_list, trajectory_data = NULL)
}
\arguments{
\item{trajectory_data}{An optional argument giving the trajectory data to supperpose.}

\item{covariates}{covariates list, that must be a list of "raster-like"
elements. It is for the moment supposed that the grid is regular and
shared by all elements of covariates}
}
\value{
a ggplot
}
\description{
Plotting the classical RSF UD
}
