% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.rhabit.R
\name{print.rhabit}
\alias{print.rhabit}
\title{Title}
\usage{
\method{print}{rhabit}(x, digits = max(3L, getOption("digits") - 3L))
}
\arguments{
\item{x}{a rhabit model fit using fit_langevin_ud}

\item{digits}{a non-null value for digits specifies the minimum number o
f significant digits to be printed in values.}
}
\value{

}
\description{
Title
}
\examples{
data(tracks)
fitted_langevin <- fit_langevin_ud(cbind(x,y) ~ grad_c1, data = tracks)
print(fitted_langevin)
}
