% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/langevinUD.R
\name{summary.rhabit}
\alias{summary.rhabit}
\title{Obtaining the summaryof a fitted  Langevin movement model}
\usage{
\method{summary}{rhabit}(rhabit_obj)
}
\arguments{
\item{rhabit_obj}{an object of class rhabit, obtained through the function langevinUD}
}
\value{
the table corresponding to the test of H0 $beta_i = 0$
}
\description{
Obtaining the summaryof a fitted  Langevin movement model
}
